UI[web]: Center cards on tablet viewport
The card displaying the text for speech input and verification
aligns to the left on a tablet viewport. This happens as one of
it's sibling in the carousel disappears with reduce width forcing
flex rule "jsutify-content: space-between" to push the element to
left. Adjusting the width to occupy the entirity of the space
recenters the card in focus.